レイアウトコンポーネント Fixes #48
