+++ navigation/AppNavigator.tsx
@@ -1,46 +1,19 @@
 import * as React from "react";
 import { View } from "react-native";
-import { createMaterialTopTabNavigator } from "@react-navigation/material-top-tabs";
 import { createBottomTabNavigator } from "@react-navigation/bottom-tabs";
 import FirstBottomTabScreen from "../screens/TimeLine";
 import SecondBottomTabScreen from "../screens/MyPage";
-import RankingChest from "../screens/RankingChest";
-import RankingBack from "../screens/RankingBack";
 import { NavigationContainer, Route } from "@react-navigation/native";
 import GroupList from "../screens/GroupList";
-import {
-  BottomTabNavigationOptions,
-  BottomTabBarOptions,
-} from "@react-navigation/bottom-tabs";
+import Header from "../components/uiParts/Header";
 import {
   MaterialCommunityIcons,
   AntDesign,
   FontAwesome5,
 } from "@expo/vector-icons";
+import RankingTabNavigator from "./RankingTabNavigator";
 
-type TabRoute = {
-  name: "Home" | "Clip";
-};
-
-const RankingTab = createMaterialTopTabNavigator();
 const BottomTab = createBottomTabNavigator();
-
-function RankingTabNavigator() {
-  return (
-    <RankingTab.Navigator
-      tabBarOptions={{
-        activeTintColor: "#594639",
-        inactiveTintColor: "rgba(89, 70, 57, 0.4)",
-        indicatorStyle: { backgroundColor: "transparent" },
-        labelStyle: { fontSize: 20 },
-        style: { marginTop: 100, backgroundColor: "#F8F4E6" },
-      }}
-    >
-      <RankingTab.Screen name="胸" component={RankingChest} />
-      <RankingTab.Screen name="背中" component={RankingBack} />
-    </RankingTab.Navigator>
-  );
-}
 
 const screenOption = ({ route }: { route: Route<string> }) => ({
   tabBarIcon: ({
@@ -87,22 +60,25 @@ const screenOption = ({ route }: { route: Route<string> }) => ({
 export default function AppNavigator() {
   return (
     <NavigationContainer>
-      <BottomTab.Navigator
-        screenOptions={screenOption}
-        tabBarOptions={{
-          activeTintColor: "#594639",
-          inactiveTintColor: "#594639",
-          style: { height: 80, backgroundColor: "#F8F4E6" },
-        }}
-      >
-        <BottomTab.Screen name="マイ記録" component={SecondBottomTabScreen} />
-        <BottomTab.Screen
-          name="タイムライン"
-          component={FirstBottomTabScreen}
-        />
-        <BottomTab.Screen name="ランキング" component={RankingTabNavigator} />
-        <BottomTab.Screen name="グループ" component={GroupList} />
-      </BottomTab.Navigator>
+      <View style={{ flex: 1 }}>
+        <Header name="My Custom Header" />
+        <BottomTab.Navigator
+          screenOptions={screenOption}
+          tabBarOptions={{
+            activeTintColor: "#594639",
+            inactiveTintColor: "#594639",
+            style: { height: 80, backgroundColor: "#F8F4E6" },
+          }}
+        >
+          <BottomTab.Screen name="マイ記録" component={SecondBottomTabScreen} />
+          <BottomTab.Screen
+            name="タイムライン"
+            component={FirstBottomTabScreen}
+          />
+          <BottomTab.Screen name="ランキング" component={RankingTabNavigator} />
+          <BottomTab.Screen name="グループ" component={GroupList} />
+        </BottomTab.Navigator>
+      </View>
     </NavigationContainer>
   );
 }
